backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0

pymongo>=4.7.0
python-dotenv>=1.0.0

passlib[bcrypt]>=1.7.4
python-jose>=3.3.0
bcrypt<5.0.0

langchain>=0.1.0
langchain-community>=0.0.1
langchain-groq>=0.1.0
langchain-text-splitters>=0.0.1

beautifulsoup4>=4.11.0
requests>=2.31.0

# FAISS and sentence-transformers for RAG features
faiss-cpu>=1.7.4
sentence-transformers>=2.2.0